backend/requirements.txt
flask
flask-cors
flask-jwt-extended
sentence-transformers
transformers
torch
numpy
pandas
scikit-learn
pymongo
python-docx
pdfminer.six

